fix remoteparticle delete

+++ classic/5.0/src/Utilities/Options.h
@@ -50,7 +50,7 @@ namespace Options {
 
     // If the distance of a particle to bunch mass larger than remotePartDel times of the rms size of the bunch in any dimension,
     // the particle will be deleted artifically to hold the accuracy of space charge calculation. The default setting of -1 stands for no deletion.
-    extern int remotePartDel;
+    extern double remotePartDel;
 
     extern double beamHaloBoundary;
 }
